Rename list methods to index in UserController and PermissionController; update routes accordingly

+++ routes/api.php
@@ -24,14 +24,14 @@ Route::middleware('auth:sanctum')->group(function () {
     Route::apiResource('/genres', GenreController::class)->except('update');
 
     // User routes
-    Route::get('/users', [UserController::class, 'list']);
+    Route::get('/users', [UserController::class, 'index']);
 
     // Admin routes
-    Route::middleware('role:admin')->prefix('/admin/users/')->group(function () {
-        Route::get('roles', [RoleController::class, 'list']);
-        Route::put('{user}/assign-role', [RoleController::class, 'assignRole']);
-        Route::get('permissions', [PermissionController::class, 'list']);
-        Route::post('{user}/give-permission', [PermissionController::class, 'givePermission']);
-        Route::delete('{user}/revoke-permission', [PermissionController::class, 'revokePermission']);
+    Route::middleware('role:admin')->group(function () {
+        Route::get('/roles', [RoleController::class, 'index']);
+        Route::put('/roles/{user}/assign-role', [RoleController::class, 'assignRole']);
+        Route::get('/roles/permissions', [PermissionController::class, 'index']);
+        Route::post('/roles/permissions/{user}/give-permission', [PermissionController::class, 'givePermission']);
+        Route::delete('/roles/permissions/{user}/revoke-permission', [PermissionController::class, 'revokePermission']);
     });
 });
